[UPDATE][FIX]BioProjectのエラーで落ちている部分を一部修正7

+++ batch/src/main/java/com/ddbj/ld/data/beans/bioproject/OrganizationName.java
@@ -40,7 +40,7 @@ public class OrganizationName {
                 case VALUE_NULL:
                     break;
                 case VALUE_NUMBER_INT:
-                    value.setContent(jsonParser.readValueAs(Integer.class).toString());
+                    value.setContent(jsonParser.readValueAs(Long.class).toString());
 
                     break;
                 case VALUE_STRING:
